+++ tests/Rust/tests/src/AsyncTests.fs
@@ -102,7 +102,6 @@ let shouldExecuteTask () =
 //     t.Join()
 //     x |> equal 2
 
-
 // [<Fact>]
 // let monitorShouldWorkWithSystemObj () =
 //     let o = new System.Object() // todo - this doesn't work, and outputs unit
@@ -175,6 +174,24 @@ let ``Lock should return result`` () =
 //     do t.Result
 //     x |> equal 2
 
+[<Fact>]
+let ``Async.Sleep works`` () =
+    let mutable s = ""
+    let a1 =
+        async {
+            do! Async.Sleep(200)
+            s <- s + "200"
+        }
+    let a2 =
+        async {
+            do! Async.Sleep(100)
+            s <- s + "100"
+        }
+    let t1 = a1 |> Async.StartAsTask
+    let t2 = a2 |> Async.StartAsTask
+    let r1 = t1.Result
+    let r2 = t2.Result
+    s |> equal "100200"
 
 // type DisposableAction(f) =
 //     interface IDisposable with
@@ -588,7 +605,6 @@ let ``Lock should return result`` () =
 //     Async.StartWithContinuations(work, (fun r -> result <- r), ignore, ignore)
 //     equal result 42
 
-
 // [<Fact>]
 // let ``Deep recursion with async doesn't cause stack overflow`` () =
 //     async {
@@ -744,7 +760,7 @@ let ``Lock should return result`` () =
 //     |> Async.StartImmediate
 
 // [<Fact>]
-// let ``Async.StartChild applys timeout`` () =
+// let ``Async.StartChild applies timeout`` () =
 //     async {
 //         let mutable x = ""
 //         let task = async {